[BpkPrice] Migrate JS to TypeScript
Convert the bpk-component-price package from JavaScript/Flow to
TypeScript. Replace prop-types and Flow Props with typed Props exports
and rename all source/test/snapshot files (.js -> .ts/.tsx). No
business-logic, JSX, or rendering changes - existing snapshots remain
valid.

- Rename index.js -> index.ts and re-export BpkPriceProps type
- Rename BpkPrice.js -> BpkPrice.tsx; type Props extends
  Omit<HTMLAttributes<HTMLDivElement>, 'className'> so native div props
  flow through ...rest while keeping className?: string | null
- Rename common-types.js -> common-types.ts; use `as const` so SIZES /
  ALIGNS export literal-typed values
- Rename test/accessibility-test/snapshot files; type the BpkPrice-test
  props variable as BpkPriceProps and cast the icon NewWindowIcon
  reference to ReactNode to preserve the existing (already-empty) icon
  span in snapshots
- Drop the now-unused @ts-expect-error suppressions on the BpkPrice and
  common-types imports in the stories file
