+++ packages/backpack-web/src/bpk-component-price/index.ts
@@ -15,24 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/* @flow strict */
 
-import { render } from '@testing-library/react';
-import { axe } from 'jest-axe';
+import BpkPrice, { type Props as BpkPriceProps } from './src/BpkPrice';
+import { SIZES, ALIGNS } from './src/common-types';
 
-import BpkPrice from './BpkPrice';
-
-describe('BpkPrice accessibility tests', () => {
-  it('should not have programmatically-detectable accessibility issues', async () => {
-    const { container } = render(
-      <BpkPrice
-        price="£1,830"
-        previousPrice="£2,000"
-        leadingText="Web only deal"
-        trailingText="per day"
-      />,
-    );
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-});
+export default BpkPrice;
+export type { BpkPriceProps };
+export { SIZES, ALIGNS };
