+++ packages/backpack-web/src/bpk-component-price/src/BpkPrice.stories.tsx
@@ -19,9 +19,7 @@
 
 import NewWindowIcon from '../../bpk-component-icon/sm/new-window';
 
-// @ts-expect-error Untyped import
 import BpkPrice from './BpkPrice';
-// @ts-expect-error Untyped import
 import { SIZES, ALIGNS } from './common-types';
 
 import type { Meta } from '@storybook/react';
